+++ python/GafferUI/ColorChooser.py
@@ -151,7 +151,7 @@ class ColorChooser( GafferUI.Widget ) :
 					)
 
 			# initial and current colour swatches
-			with GafferUI.ListContainer( GafferUI.ListContainer.Orientation.Horizontal, spacing = 4 ) :
+			with GafferUI.ListContainer( GafferUI.ListContainer.Orientation.Horizontal, spacing = 4 ) as self.__swatchRow :
 
 				self.__initialColorSwatch = GafferUI.ColorSwatch( color, useDisplayTransform = useDisplayTransform, parenting = { "expand" : True } )
 				self.__initialColorSwatch._qtWidget().setFixedHeight( 40 )
@@ -182,6 +182,14 @@ class ColorChooser( GafferUI.Widget ) :
 	def getColor( self ) :
 
 		return self.__color
+
+	def setSwatchesVisible( self, visible ) :
+
+		self.__swatchRow.setVisible( False )
+
+	def getSwatchesVisible( self ) :
+
+		return self.__swatchRow.getVisible()
 
 	## A signal emitted whenever the color is changed. Slots should
 	# have the signature slot( ColorChooser, reason ). The reason